groq-chat/requirements.txt
groq==0.9.0
chainlit==1.1.402
exa-py==1.0.17
python-dotenv==1.0.1